Move test data into tests

+++ src/main/resources/db/migration/V1__init.sql
@@ -3,6 +3,6 @@ CREATE TABLE item
     barcode text  NOT NULL primary key,
     name    text  NOT NULL,
     unit    text  NOT NULL,
-    price   money NOT NULL,
+    price   float NOT NULL,
     type    text   NOT NULL
 );
